--Move Attributes to their own directory and namespace.
