+++ src/esp/bindings/AttributesManagersBindings.cpp
@@ -8,7 +8,7 @@
 #include <Magnum/Magnum.h>
 #include <Magnum/PythonBindings.h>
 
-#include "esp/assets/Attributes.h"
+#include "esp/assets/attributes/Attributes.h"
 
 #include "esp/assets/managers/AttributesManagerBase.h"
